Add Wasm memory64 support

+++ crates/wasm-encoder/src/lib.rs
@@ -129,8 +129,8 @@ impl Encode for str {
 
 impl Encode for usize {
     fn encode(&self, sink: &mut Vec<u8>) {
-        assert!(*self <= u32::max_value() as usize);
-        (*self as u32).encode(sink)
+        assert!(*self <= u64::max_value() as usize);
+        (*self as u64).encode(sink)
     }
 }
